見た目の調整とパフォーマンスの修正

+++ Soul.pde
@@ -1,6 +1,6 @@
 class Soul{
     PVector location;
-    float size;
+    int size;
     int life;
     boolean isDead = false;
